DuckyScript
This Script Created a WiFi Hotspot on Victim's PC with Low FireWall Protection with Key 1234567890 and then grabs saved WiFi Passwords And sends it to Gmail and Launching a fake Win10 Update Screen. This will Make the Victim Think That his PC is updating so he will probably wait for hours because this fake update is never going to complete.